Migrate textarea from /docs/components/ to /components/
