+++ site/ui/e2e/textarea.spec.ts
@@ -1,8 +1,8 @@
 import { test, expect } from '@playwright/test'
 
-test.describe('Textarea Documentation Page', () => {
+test.describe('Textarea Reference Page', () => {
   test.beforeEach(async ({ page }) => {
-    await page.goto('/docs/components/textarea')
+    await page.goto('/components/textarea')
   })
 
   test.describe('Textarea Rendering', () => {
@@ -13,27 +13,12 @@ test.describe('Textarea Documentation Page', () => {
 
     test('has multiple textarea examples', async ({ page }) => {
       const textareas = page.locator('textarea[data-slot="textarea"]')
-      // Should have textareas on the page (preview + examples)
+      // Should have textareas on the page (playground + usage + examples)
       expect(await textareas.count()).toBeGreaterThanOrEqual(4)
     })
   })
 
-  test.describe('Basic', () => {
-    test('displays basic example', async ({ page }) => {
-      await expect(page.locator('h3:has-text("Basic")')).toBeVisible()
-    })
-
-    test('has placeholder text', async ({ page }) => {
-      const textareas = page.locator('textarea[data-slot="textarea"]')
-      await expect(textareas.first()).toHaveAttribute('placeholder', 'Type your message here.')
-    })
-  })
-
   test.describe('Disabled', () => {
-    test('displays disabled example', async ({ page }) => {
-      await expect(page.locator('h3:has-text("Disabled")')).toBeVisible()
-    })
-
     test('has disabled textarea', async ({ page }) => {
       const disabledTextarea = page.locator('textarea[data-slot="textarea"][disabled]')
       await expect(disabledTextarea.first()).toBeVisible()
@@ -41,10 +26,6 @@ test.describe('Textarea Documentation Page', () => {
   })
 
   test.describe('Value Binding', () => {
-    test('displays value binding example', async ({ page }) => {
-      await expect(page.locator('h3:has-text("Value Binding")')).toBeVisible()
-    })
-
     test('updates character count on input', async ({ page }) => {
       const section = page.locator('[bf-s^="TextareaBindingDemo_"]:not([data-slot])').first()
       await expect(section).toBeVisible()
@@ -61,5 +42,4 @@ test.describe('Textarea Documentation Page', () => {
       await expect(charCount).toContainText('5')
     })
   })
-
 })
