+++ site/ui/pages/components/textarea.tsx
@@ -6,6 +6,7 @@
  */
 
 import { Textarea } from '@/components/ui/textarea'
+import { TextareaBindingDemo } from '@/components/textarea-demo'
 import { TextareaPlayground } from '@/components/textarea-playground'
 import {
   DocPage,
@@ -23,6 +24,9 @@ const tocItems: TocItem[] = [
   { id: 'preview', title: 'Preview' },
   { id: 'installation', title: 'Installation' },
   { id: 'usage', title: 'Usage' },
+  { id: 'examples', title: 'Examples' },
+  { id: 'disabled', title: 'Disabled', branch: 'start' },
+  { id: 'value-binding', title: 'Value Binding', branch: 'end' },
   { id: 'api-reference', title: 'API Reference' },
 ]
 
@@ -35,6 +39,41 @@ function TextareaDemo() {
       <Textarea disabled placeholder="Disabled textarea" />
       <Textarea error placeholder="Error state" />
       <Textarea rows={6} placeholder="With explicit rows" />
+    </div>
+  )
+}`
+
+const disabledCode = `"use client"
+
+import { Textarea } from '@/components/ui/textarea'
+
+function TextareaDisabled() {
+  return (
+    <div className="flex flex-col gap-2 max-w-sm">
+      <Textarea disabled placeholder="Disabled textarea" />
+      <Textarea readOnly value="Read-only content" />
+    </div>
+  )
+}`
+
+const bindingCode = `"use client"
+
+import { createSignal } from '@barefootjs/dom'
+import { Textarea } from '@/components/ui/textarea'
+
+function TextareaBinding() {
+  const [value, setValue] = createSignal('')
+
+  return (
+    <div className="max-w-sm space-y-2">
+      <Textarea
+        value={value()}
+        onInput={(e) => setValue(e.target.value)}
+        placeholder="Type your message here."
+      />
+      <p className="text-sm text-muted-foreground">
+        {value().length} characters
+      </p>
     </div>
   )
 }`
@@ -123,6 +162,24 @@ export function TextareaRefPage() {
               <Textarea rows={6} placeholder="With explicit rows" />
             </div>
           </Example>
+        </Section>
+
+        {/* Examples */}
+        <Section id="examples" title="Examples">
+          <div className="space-y-8">
+            <Example title="Disabled" code={disabledCode}>
+              <div className="flex flex-col gap-2 max-w-sm">
+                <Textarea disabled placeholder="Disabled textarea" />
+                <Textarea readonly value="Read-only content" />
+              </div>
+            </Example>
+
+            <Example title="Value Binding" code={bindingCode}>
+              <div className="max-w-sm">
+                <TextareaBindingDemo />
+              </div>
+            </Example>
+          </div>
         </Section>
 
         {/* API Reference */}
